+++ Tutorial15_GraphicsEngine_RenderSystem_RAII_Exceptions/VertexBuffer.cpp
@@ -13,7 +13,7 @@
 #include <exception>
 
 VertexBuffer::VertexBuffer(void* list_vertices,UINT size_vertex,UINT size_list,
-	void*shader_byte_code,UINT size_byte_shader,
+	void*shader_byte_code,size_t size_byte_shader,
 	RenderSystem * system) : m_system(system) ,m_layout(0),m_buffer(0)
 {
 	D3D11_BUFFER_DESC buff_desc = {};
