+++ Tutorial15_GraphicsEngine_RenderSystem_RAII_Exceptions/VertexBuffer.h
@@ -15,7 +15,7 @@
 class VertexBuffer
 {
 public:
-	VertexBuffer(void* list_vertices, UINT size_vertex, UINT size_list, void*shader_byte_code, UINT size_byte_shader,
+	VertexBuffer(void* list_vertices, UINT size_vertex, UINT size_list, void*shader_byte_code, size_t size_byte_shader,
 		RenderSystem * system);
 	UINT getSizeVertexList();
 	~VertexBuffer();
